FEATURE: More user friendly login animation UI +Mobile tweak
It seems to me that any login UI animation is best
received on the actual element a user interacts with. Hence the decision to put the reactstrap spinner in the
next button.  Also added a no pointer event so the button
is disabled while login process is underway. Lastly added
some minor improvements to mobile.