FEATURE:Complete Layout for Statistics component
Important:push path in login action changed to dashboard.You can easily change this back if you want to flow from login to the component you are working on.e
Visually we are about 80% of the way there.  Will need 3 other graph components for the weekly, monthly, and yearly stats per product.  We will also need a section for basic retailer and organization counts.